+++ extensions/Worker.Extensions.Http.AspNetCore/src/HttpContextConverter.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.Functions.Worker.Converters;
@@ -10,7 +9,7 @@ using Microsoft.Azure.Functions.Worker.Http;
 namespace Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore
 {
     /// <summary>
-    /// Converter to bind HttpRequest type parameters.
+    /// Converter to bind <see cref="HttpContext"/>, <see cref="HttpRequest"/> and <see cref="HttpRequestData"/> type parameters.
     /// </summary>
     internal class HttpContextConverter : IInputConverter
     {
@@ -18,7 +17,11 @@ namespace Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore
         {
             object? target = null;
 
-            if (context.TargetType == typeof(HttpRequest)
+            if (context.TargetType == typeof(HttpContext))
+            {
+                target = context.FunctionContext.GetHttpContext();
+            }
+            else if (context.TargetType == typeof(HttpRequest)
                 && context.FunctionContext.TryGetRequest(out var request))
             {
                 target = request;